+++ ChangeLog.md
@@ -24,7 +24,7 @@ Full details of the variables available for each noted event, and VoiceAttack in
     * `Carrier pads locked` updated to include `carrierID` and `carrierType` fields.
     * `Carrier purchased` updated to include `carrierID` and `carrierType` fields.
     * `Carrier service changed` updated to include `carrierID` and `carrierType` fields.
-    * `Carrier stats` updated to include `carrierID` and `carrierType` fields.`
+    * `Carrier stats` updated to include `carrierID`, `carrierType`, `jumpRange`, `jumpRangeMax`, and `pendingDecommission` fields.
     * `Squadron rank` updated to include new Squadron rank objects and the new `squadronID` integer.
     * `Squadron startup` updated to include a new Squadron rank object and the new `squadronID` integer.
     * `Squadron status` updated to include the new `squadronID` integer.